// angular-frontend/src/app/interfaces/driver.ts
export class Driver {
    id: number;
  vehicleNumber: string;
  driverName: string;
  vehicleType: string;
  capacity: number;
  costPerSlot: number;
  vehicleStatus: string;
  // date: string;
  slot1: string;
  slot2: string;
  slot3: string;
  bookingId:number;
  companyName: string;
  slot: string;
  retailerId: number;
  requestStatus: string;
  remainingCapacity: number;
  password: string;
}